if cannot find ci scripts, report build failed

+++ autobuild.sh
@@ -7,9 +7,14 @@ while [ "$(ls -A out/pending)" ]; do
     for file in out/pending/*; do
         set -m
         ref=`basename $file`
+        # clean up build/
         ./clean-dir.sh $ref
         # run build
-        $CI_PATH/run-build.sh $ref | tee out/log
+        if [ -f $CI_PATH/run-build.sh ]; then
+            $CI_PATH/run-build.sh $ref | tee out/log
+        else
+            echo "cannot find $CI_PATH/run-build.sh" > out/fail/$ref
+        fi
         # make archive
         if [ -f out/pass/$ref ]; then
             $CI_PATH/archive-bin.sh $ref
